requirements.txt
pyserial
requests
bs4
zpl